+++ Source/Core/Visualization/Event/EventHandler.h
@@ -3,18 +3,8 @@
  *  @file   EventHandler.h
  *  @author Naohisa Sakamoto
  */
-/*----------------------------------------------------------------------------
- *
- *  Copyright (c) Visualization Laboratory, Kyoto University.
- *  All rights reserved.
- *  See http://www.viz.media.kyoto-u.ac.jp/kvs/copyright/ for details.
- *
- *  $Id: EventHandler.h 1325 2012-10-04 10:34:52Z naohisa.sakamoto@gmail.com $
- */
 /*****************************************************************************/
-#ifndef KVS__EVENT_HANDLER_H_INCLUDE
-#define KVS__EVENT_HANDLER_H_INCLUDE
-
+#pragma once
 #include <vector>
 #include <cstddef>
 #include <string>
@@ -33,14 +23,12 @@ class EventBase;
 /*===========================================================================*/
 class EventHandler
 {
-protected:
-
+private:
     std::vector<kvs::EventListener*> m_listeners; ///< list of the event listener
 
 public:
-
-    EventHandler();
-    virtual ~EventHandler();
+    EventHandler() {}
+    virtual ~EventHandler() {}
 
     void attach( kvs::EventListener* listener );
     void detach( const kvs::EventListener* listener );
@@ -50,5 +38,3 @@ public:
 };
 
 } // end of namespace kvs
-
-#endif // KVS__EVENT_HANDLER_H_INCLUDE