+++ Source/Core/Visualization/Event/InitializeEvent.cpp
@@ -3,16 +3,9 @@
  *  @file   InitializeEvent.cpp
  *  @author Naohisa Sakamoto
  */
-/*----------------------------------------------------------------------------
- *
- *  Copyright (c) Visualization Laboratory, Kyoto University.
- *  All rights reserved.
- *  See http://www.viz.media.kyoto-u.ac.jp/kvs/copyright/ for details.
- *
- *  $Id$
- */
 /*****************************************************************************/
 #include "InitializeEvent.h"
+
 
 namespace kvs
 {