+++ Source/Core/Visualization/Event/KeyEvent.h
@@ -24,6 +24,7 @@ private:
     int m_x; ///< mouse cursol position x
     int m_y; ///< mouse cursol position y
     int m_modifiers; ///< modifier keys
+    int m_action; ///< key action (pressed, released, or repeated)
 
 public:
     KeyEvent();
@@ -35,11 +36,13 @@ public:
     int x() const;
     int y() const;
     int modifiers() const;
+    int action() const { return m_action; }
     int type() const;
 
     void setKey( int key );
     void setPosition( int x, int y );
     void setModifiers( int modifiers );
+    void setAction( int action ) { m_action = action; }
 };
 
 } // end of namespace kvs